requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-wtf>=1.2
pdfplumber>=0.10
pypdf>=4.0
google-auth>=2.0
google-api-python-client>=2.0
werkzeug>=3.0
python-dotenv>=1.0
gunicorn>=21.2
psycopg2-binary>=2.9
